+++ tests/50-mailtest.py
@@ -20,12 +20,14 @@
 
 # Test mail sending
 
+from datetime import datetime, timezone
 import os
 from pathlib import Path
+import pygit2 as git
 import tempfile
 import time
-from datetime import datetime, timezone
 import threading
+import subprocess
 import sys
 sys.path.append('.')
 
@@ -37,7 +39,7 @@ def mailtest():
   tmpdir = tempfile.TemporaryDirectory()
   zeitgitter.config.get_args(args=[
     '-c', 'tests/mailtest.conf',
-    '--debug', 2,
+    '--debug', '2',
     '--gnupg-home',
     str(Path(os.path.dirname(os.path.realpath(__file__)),
                      'gnupg')),
@@ -115,7 +117,14 @@ fa94ffe675454658bd11219693d60844b995a74d
                    hour=16, minute=55, second=0,
                    tzinfo=timezone.utc).timestamp()
   os.utime(p, times=(ftime, ftime))
-  zeitgitter.mail.receive_async()
+  subprocess.run(['git', 'init'],
+          cwd=zeitgitter.config.arg.repository).check_returncode()
+  subprocess.run(['git', 'add', 'hashes.log'],
+          cwd=zeitgitter.config.arg.repository).check_returncode()
+  subprocess.run(['git', 'commit', '-m', "First commit"],
+          cwd=zeitgitter.config.arg.repository).check_returncode()
+  repo = git.Repository(zeitgitter.config.arg.repository)
+  zeitgitter.mail.wait_for_receive(repo, repo.head, p)
 
 if os.path.isfile('tests/mailtest.conf'):
     mailtest()