+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Fixed
 - Correctly handles IMAP `IDLE` responses other than `EXISTS` (especially
   Dovecot's `* OK still here`)
+- End line in stamper mails may now also be in last line.
 
 ## Changed
 - Split into client (git-timestamp) and server (zeitgitterd).